Use HDF5 type for QUAL
As a side effect of a previous update, a better type definition is
needed is QUAL.

+++ dsm2/src/qual/qual_hdf_output.f90
@@ -301,8 +301,8 @@
       integer(HSIZE_T), dimension(7) :: cg_data_dims
       integer(HID_T) :: memspace ! memspace identifier
 
-      integer, parameter :: label_len = 12
-      integer, parameter :: name_len=32
+      integer(SIZE_T), parameter :: label_len = 12
+      integer(SIZE_T), parameter :: name_len=32
       character(LEN=label_len),dimension(2) :: chan_location &
                                          = (/"upstream","downstream"/)
       character(LEN=name_len),dimension(:), allocatable :: names
